requirements.txt
cycler==0.10.0
decorator==5.0.7
future==0.18.2
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.1.3
matplotlib-inline==0.1.7
networkx==2.4
numpy==1.22.0
pandas>=0.24.2
plotly==4.14.3
pyparsing==2.4.7
pyquaternion==0.9.9
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.9.1
six==1.15.0
torch>=1.7
torchplot==0.2.0
control==0.9.4
roboticstoolbox-python==1.1.0
spatialgeometry==1.1.0
spatialmath-python==1.1.9
swift-sim==1.1.0